refactor: move problem_report_reason to problem_report
